fix(ui): don't show 'copy invite code' message if we don't know which conversation the invite code should be for (i.e. localstorage was deleted)
